% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{address_neighbors}
\alias{address_neighbors}
\title{Get directly connected addresses of an address}
\usage{
address_neighbors(address, direction = c("in", "out"),
  currency = supported_currencies(), pagesize = 1000L,
  api = get_api(), token = get_token())
}
\arguments{
\item{address}{A character scalar to specify a cryptocurrency address.}

\item{direction}{A character scalar to specify the direction ("in", "out")}

\item{currency}{A character scalar to specify the currency ticker symbol.}

\item{pagesize}{An integer scalar to specify the size of paginated results.}

\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Get directly connected addresses of an address
}
\examples{
\dontrun{
set_token("GRAPHSENSE_API_TOKEN")
address_neighbors("16SLJL6RCqHjySsKEdaEnNRWTrMhv6S8Z5")
}
}
