% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{authenticate}
\alias{authenticate}
\title{Authentication function for GraphSense API}
\usage{
authenticate(username, password, api = get_api())
}
\arguments{
\item{username}{A character scalar to specify the username.}

\item{password}{A character scalar to specify the password.}

\item{api}{GraphSense API URL.}
}
\description{
Log into GraphSense API and retrieve authorization token
}
