% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{block_txs}
\alias{block_txs}
\title{Get all transaction in a block}
\usage{
block_txs(block, currency = supported_currencies(), api = get_api(),
  token = get_token())
}
\arguments{
\item{block}{An integer scalar to specify a block number.}

\item{currency}{A character scalar to specify the currency ticker symbol.}

\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Get all transaction in a block
}
\examples{
\dontrun{
set_token("GRAPHSENSE_API_TOKEN")
block_txs(100000)
}
}
