% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Get basic cluster information}
\usage{
cluster(cluster, currency = supported_currencies(), api = get_api(),
  token = get_token())
}
\arguments{
\item{cluster}{An integer scalar to specify a cluster ID.}

\item{currency}{A character scalar to specify the currency ticker symbol.}

\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Get basic cluster information
}
\examples{
\dontrun{
set_token("GRAPHSENSE_API_TOKEN")
cluster(17642139)
}
}
