% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_addresses}
\alias{cluster_addresses}
\title{Get basic information of all addresses in a cluster}
\usage{
cluster_addresses(cluster, currency = supported_currencies(),
  pagesize = 1000L, api = get_api(), token = get_token())
}
\arguments{
\item{cluster}{An integer scalar to specify a cluster ID.}

\item{currency}{A character scalar to specify the currency ticker symbol.}

\item{pagesize}{An integer scalar to specify the size of paginated results.}

\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Get basic information of all addresses in a cluster
}
\examples{
\dontrun{
set_token("GRAPHSENSE_API_TOKEN")
cluster(17642139)
}
}
