% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{exchange_rate}
\alias{exchange_rate}
\title{Get exchange rate information}
\usage{
exchange_rate(height, currency = supported_currencies(),
  api = get_api(), token = get_token())
}
\arguments{
\item{height}{An integer scalar to specify a block number.}

\item{currency}{A character scalar to specify the currency ticker symbol.}

\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Get exchange rate information
}
\examples{
\dontrun{
exchange_rate(100000)
}
}
