% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{is_known_address}
\alias{is_known_address}
\title{Check if an address is known by GraphSense}
\usage{
is_known_address(address, currency = supported_currencies(),
  api = get_api(), token = get_token())
}
\arguments{
\item{address}{A character scalar to specify a cryptocurrency address.}

\item{currency}{A character scalar to specify the currency ticker symbol.}

\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Check if an address is known by GraphSense
}
