% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block.R
\name{is_known_block}
\alias{is_known_block}
\title{Check if a block is included in GraphSense}
\usage{
is_known_block(block, currency = supported_currencies(),
  api = get_api(), token = get_token())
}
\arguments{
\item{block}{An integer scalar to specify a cryptocurrency address.}

\item{currency}{A character scalar to specify the currency ticker symbol.}

\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Check if a block is included in GraphSense
}
