% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{is_known_cluster}
\alias{is_known_cluster}
\title{Check if a cluster is present in GraphSense}
\usage{
is_known_cluster(cluster, currency = supported_currencies(),
  api = get_api(), token = get_token())
}
\arguments{
\item{cluster}{An integer scalar to specify a cluster ID.}

\item{currency}{A character scalar to specify the currency ticker symbol.}

\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Check if a cluster is present in GraphSense
}
