% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{statistics}
\alias{statistics}
\title{Summary statistics of address/cluster graph for all currencies}
\usage{
statistics(api = get_api(), token = get_token())
}
\arguments{
\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Summary statistics of address/cluster graph for all currencies
}
