% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx.R
\name{tx}
\alias{tx}
\title{Get basic transaction information}
\usage{
tx(tx_hash, currency = supported_currencies(), api = get_api(),
  token = get_token())
}
\arguments{
\item{tx_hash}{A character scalar to specify a transaction hash.}

\item{currency}{A character scalar to specify the currency ticker symbol.}

\item{api}{GraphSense API URL.}

\item{token}{GraphSense API token.}
}
\description{
Get basic transaction information
}
\examples{
\dontrun{
set_token("GRAPHSENSE_API_TOKEN")
tx("25b8d2942a5bdf36bfbae9567edc8744367bf5c2e61515ec0fa79af531c73aca")
}
}
